requirements.txt
aniso8601==9.0.1
asgiref==3.6.0
async-generator==1.10
attrs==22.2.0
autocomplete==0.0.104
beautifulsoup4==4.11.2
blinker==1.4
bottle==0.12.24
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
ConfigArgParse==1.5.3
contourpy==1.0.7
cryptography==3.4.8
cycler==0.11.0
distro==1.7.0
Django==4.1.6
django-adminlte-ui==1.7.2
django-cleanup==6.0.0
django-filter==22.1
django-treebeard==4.6.1
djangorestframework==3.14.0
et-xmlfile==1.1.0
Flask==2.2.3
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
fonttools==4.38.0
gevent==22.10.2
geventhttpclient==2.0.8
google==3.0.0
gpt==0.2
graphene==3.2.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==2.0.1
h11==0.14.0
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
inexactsearch==1.0.2
iniconfig==2.0.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.2.0
keyring==23.5.0
kiwisolver==1.4.4
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
locust==2.14.2
MarkupSafe==2.1.2
matplotlib==3.7.0
more-itertools==8.10.0
msgpack==1.0.4
NlpToolkit-Corpus==1.0.25
NlpToolkit-DataStructure==1.0.14
NlpToolkit-Dictionary==1.0.34
NlpToolkit-Math==1.0.18
NlpToolkit-MorphologicalAnalysis==1.0.47
NlpToolkit-NGram==1.0.19
NlpToolkit-Sampling==1.0.7
NlpToolkit-SpellChecker==1.0.24
nltk==3.8.1
numpy==1.24.2
oauthlib==3.2.0
opencv-python==4.7.0.68
openpyxl==3.1.0
outcome==1.2.0
packaging==23.0
pandas==1.5.3
Pillow==9.4.0
playwright==1.30.0
pluggy==1.0.0
promise==2.3
psutil==5.9.4
psycopg2==2.9.5
pycparser==2.21
pyee==9.0.4
PyJWT==2.3.0
pyparsing==2.4.7
PyPDF2==3.0.1
PySocks==1.7.1
pyspellchecker==0.7.1
pytesseract==0.3.10
pytest==7.2.1
pytest-base-url==2.0.0
pytest-playwright==0.3.0
python-dateutil==2.8.2
python-slugify==8.0.0
pytz==2022.7.1
pywin32==305
pywin32-ctypes==0.2.0
PyYAML==5.4.1
pyzmq==25.0.0
regex==2022.10.31
requests==2.28.2
roundrobin==0.0.4
scikit-learn==1.2.1
scipy==1.10.0
SecretStorage==3.3.1
selenium==4.8.0
selenium-stealth==1.0.6
silpa-common==0.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soundex==1.1.3
soupsieve==2.4
spellchecker==0.4
spellchecker-ml==0.1.4
sqlparse==0.4.3
tabula-py==2.7.0
tesseract==0.1.3
text-unidecode==1.3
textblob==0.17.1
threadpoolctl==3.1.0
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
tzdata==2022.7
urllib3==1.26.14
vocab==0.0.5
wadllib==1.3.6
Werkzeug==2.2.3
wsproto==1.2.0
xlrd==2.0.1
xlwt==1.3.0
YandexSpeller==1.0.0
zipp==1.0.0
zope.event==4.6
zope.interface==5.5.2
